# Conflicts:
#	src/main/java/com/example/lineofduty/common/exception/ErrorMessage.java
#	src/main/java/com/example/lineofduty/domain/enlistmentSchedule/service/EnlistmentScheduleService.java
#	src/main/java/com/example/lineofduty/domain/order/Order.java
#	src/main/java/com/example/lineofduty/domain/order/service/OrderService.java
#	src/main/java/com/example/lineofduty/domain/user/repository/UserRepository.java

+++ src/main/java/com/example/lineofduty/domain/product/service/ProductService.java
@@ -3,15 +3,10 @@ package com.example.lineofduty.domain.product.service;
 import com.example.lineofduty.common.exception.CustomException;
 import com.example.lineofduty.common.exception.ErrorMessage;
 import com.example.lineofduty.common.model.enums.ApplicationStatus;
-import com.example.lineofduty.domain.product.dto.request.ProductCreateRequest;
-import com.example.lineofduty.domain.product.dto.request.ProductUpdateRequest;
-import com.example.lineofduty.domain.product.dto.response.ProductCreateResponse;
-import com.example.lineofduty.domain.product.dto.response.ProductGetAllResponse;
-import com.example.lineofduty.domain.product.dto.response.ProductGetOneResponse;
-import com.example.lineofduty.domain.product.dto.response.ProductUpdateResponse;
+import com.example.lineofduty.domain.product.dto.request.ProductRequest;
+import com.example.lineofduty.domain.product.dto.response.ProductResponse;
 import com.example.lineofduty.domain.product.repository.ProductRepository;
-import com.example.lineofduty.domain.user.repository.UserRepository;
-import com.example.lineofduty.entity.Product;
+import com.example.lineofduty.domain.product.Product;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -26,47 +21,61 @@ public class ProductService {
 
     // 상품 등록
     @Transactional
-    public ProductCreateResponse createProduct(ProductCreateRequest request) {
+    public ProductResponse createProduct(ProductRequest request) {
+
+        if (request.getName() == null || request.getDescription() == null) {
+            throw new CustomException(ErrorMessage.MISSING_PRODUCT_NAME_OR_DESCRIPTION);
+        }
+
+        if (request.getPrice() <= 0) {
+            throw new CustomException(ErrorMessage.INVALID_PRICE);
+        }
+
+        if (request.getStock() <= 0) {
+            throw new CustomException(ErrorMessage.INVALID_STOCK);
+        }
+
         Product product = new Product(request.getName(), request.getDescription(), request.getPrice(), request.getStock(), ApplicationStatus.ProductStatus.ON_SALE);
         Product savedProduct = productRepository.save(product);
 
-        return ProductCreateResponse.from(savedProduct);
+        return ProductResponse.from(savedProduct);
     }
 
     // 상품 단건 조회
     @Transactional(readOnly = true)
-    public ProductGetOneResponse getProduct(Long productId) {
+    public ProductResponse getProduct(Long productId) {
         Product product = productRepository.findById(productId)
                 .orElseThrow(() -> new CustomException(ErrorMessage.PRODUCT_NOT_FOUND));
 
-        return ProductGetOneResponse.from(product);
+        return ProductResponse.from(product);
     }
 
     // 상품 목록 조회
     @Transactional(readOnly = true)
-    public Page<ProductGetAllResponse> getProductList(Pageable pageable) {
+    public Page<ProductResponse> getProductList(Pageable pageable) {
         Page<Product> products = productRepository.findAll(pageable);
 
-        return products.map(ProductGetAllResponse::from);
+        return products.map(ProductResponse::from);
     }
 
     // 상품 수정
     @Transactional
-    public ProductUpdateResponse updateProduct(ProductUpdateRequest request, Long productId) {
+    public ProductResponse updateProduct(ProductRequest request, Long productId) {
         Product product = productRepository.findById(productId)
                 .orElseThrow(() -> new CustomException(ErrorMessage.PRODUCT_NOT_FOUND));
 
         product.update(request);
         productRepository.saveAndFlush(product);
 
-        return ProductUpdateResponse.from(product);
+        return ProductResponse.from(product);
     }
 
     // 상품 삭제
     @Transactional
     public void deleteProduct(Long productId) {
-        Product product = productRepository.findById(productId)
-                .orElseThrow(() -> new CustomException(ErrorMessage.PRODUCT_NOT_FOUND));
+        if (!productRepository.existsById(productId)) {
+            throw new CustomException(ErrorMessage.PRODUCT_NOT_FOUND);
+        }
 
         productRepository.deleteById(productId);
     }
